+++ ktlint-ruleset-experimental/src/main/kotlin/com/github/shyiko/ktlint/ruleset/experimental/ExperimentalRuleSetProvider.kt
@@ -7,6 +7,7 @@ class ExperimentalRuleSetProvider : RuleSetProvider {
 
     override fun get(): RuleSet = RuleSet(
         "experimental",
+        ImportOrderingRule(),
         IndentationRule()
     )
 }
