+++ ktlint-ruleset-standard/src/main/kotlin/com/github/shyiko/ktlint/ruleset/standard/StandardRuleSetProvider.kt
@@ -18,8 +18,6 @@ class StandardRuleSetProvider : RuleSetProvider {
         // disabled until auto-correct is working properly
         // (e.g. try formatting "if (true)\n    return { _ ->\n        _\n}")
         // MultiLineIfElseRule(),
-        // disabled until it's clear how to reconcile difference in Intellij & Android Studio import layout
-        // ImportOrderingRule(),
         IndentationRule(),
         MaxLineLengthRule(),
         ModifierOrderRule(),
